Add MACHINE_NETWORK

+++ scripts/create-cluster.sh
@@ -60,6 +60,13 @@ then
 	exit 1
 fi
 
+MACHINE_NETWORK=$(openstack --os-cloud=${CLOUD} network show "${SUBNET_ID}" --format shell | grep ^cidr)
+if [ -z "${MACHINE_NETWORK}" ]
+then
+	echo "Error: MACHINE_NETWORK is empty!"
+	exit 1
+fi
+
 if [ -d ${CLUSTER_DIR} ]
 then
 	/bin/rm -rf ${CLUSTER_DIR}
@@ -116,6 +123,11 @@ then
 	exit 1
 fi
 
+if !getent ahostsv4 api.${CLUSTER_NAME}.${BASEDOMAIN} > /dev/null 2>&1
+then
+	echo "Error: Cannot resolve api.${CLUSTER_NAME}.${BASEDOMAIN}"
+	exit 1
+fi
 for (( TRIES=0; TRIES<=60; TRIES++ ))
 do
 	set +e
@@ -127,15 +139,10 @@ do
 	then
 		break
 	else
-		echo "Warning: VIP_API (${VIP_API}) is not the same as IP (${IP})"
+		echo "Warning: VIP_API (${VIP_API}) is not the same as IP (${IP}), sleeping..."
 	fi
 	sleep 15s
 done
-if !getent ahostsv4 api.${CLUSTER_NAME}.${BASEDOMAIN} > /dev/null 2>&1
-then
-	echo "Error: Cannot resolve api.${CLUSTER_NAME}.${BASEDOMAIN}"
-	exit 1
-fi
 IP=$(getent ahostsv4 api.${CLUSTER_NAME}.${BASEDOMAIN} 2>/dev/null | grep STREAM | cut -f1 -d' ')
 if [ "${IP}" != "${VIP_API}" ]
 then
@@ -175,7 +182,7 @@ networking:
   - cidr: 10.116.0.0/14
     hostPrefix: 23
   machineNetwork:
-  - cidr: 10.130.32.0/20
+  - cidr: ${MACHINE_NETWORK}
   networkType: OVNKubernetes
   serviceNetwork:
   - 172.30.0.0/16
